# frontend/scripts/openapi.yaml
openapi: 3.0.3
info:
  title: ''
  version: 0.0.0
paths:
...
  /users/me/comparisons/{video_id}/:
    get:
      operationId: users_me_comparisons_list_filtered
      description: List all comparisons made by the logged user.
      parameters:
      - name: limit
        required: false
        in: query
        description: Number of results to return per page.
        schema:
          type: integer
      - name: offset
        required: false
        in: query
        description: The initial index from which to return the results.
        schema:
          type: integer
      - in: path
        name: video_id
        schema:
          type: string
        required: true
      tags:
      - users
      security:
      - oauth2:
        - read write groups
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedComparisonList'
          description: ''
...
components:
  schemas:
...
    Comparison:
      type: object
      description: |-
        A comparison serializer used for the operations create, retrieve and list.

        Given the context determined by the view, it will represent the comparison
        in the reverse order.

        Use `ComparisonUpdateSerializer` for the update operation.
      properties:
        video_a:
          $ref: '#/components/schemas/VideoReadOnly'
        video_b:
          $ref: '#/components/schemas/VideoReadOnly'
        criteria_scores:
          type: array
          items:
            $ref: '#/components/schemas/ComparisonCriteriaScore'
        duration_ms:
          type: number
          format: float
          nullable: true
          description: Time it took to rate the videos (in milliseconds)
      required:
      - criteria_scores
      - video_a
      - video_b
    ComparisonCriteriaScore:
      type: object
      properties:
        criteria:
          type: string
          description: Name of the criteria
          maxLength: 32
        score:
          type: number
          format: float
          description: Score for the given comparison
        weight:
          type: number
          format: float
          description: Weight of the comparison
      required:
      - criteria
      - score
...
    PaginatedComparisonList:
      type: object
      properties:
        count:
          type: integer
          example: 123
        next:
          type: string
          nullable: true
          format: uri
          example: http://api.example.org/accounts/?offset=400&limit=100
        previous:
          type: string
          nullable: true
          format: uri
          example: http://api.example.org/accounts/?offset=200&limit=100
        results:
          type: array
          items:
            $ref: '#/components/schemas/Comparison'
...
    VideoReadOnly:
      type: object
      description: |-
        A video serializer without all the Video model auto-validations.

        Used by ModelSerializer(s) having one or more nested relations with Video,
        and having the constraint of not creating new videos in database when
        creating new objects.

        ex: adding a new Comparison shouldn't create new Video in the database
      properties:
        video_id:
          type: string
          maxLength: 20
      required:
      - video_id
...
  securitySchemes:
    oauth2:
      type: oauth2
      flows:
        password:
          tokenUrl: http://127.0.0.1:8000/o/token/
          refreshUrl: http://127.0.0.1:8000/o/token/
          scopes:
            openid: OpenID Connect scope
            read: Read scope
            write: Write scope
            groups: Access to your groups